# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: NOMOS API
  version: 0.1.0
  description: |
    에이전트 협업 조율 서버. 수동 테스트 절차는 `docs/manual-test.md`를 함께 보라.

    **토큰이 두 종류다.** 오른쪽 위 Authorize를 누르면 칸이 두 개 뜬다.

    | 칸 | 어떤 토큰 | 받는 법 |
    |---|---|---|
    | `userToken` | 사람 토큰 (`kind: user`) | `POST /auth/login` — 시드 계정 `rep` / `manual-test-1234` |
    | `agentToken` | 에이전트 토큰 (`kind: agent`) | `npm run seed` 출력의 **be-laptop access token**, 또는 `POST /agents/token/refresh` |

    서로의 API를 부를 수 없다. 사람 토큰으로 태스크 API를 부르면 401이다.

    access token은 **1시간**이다. 만료되면 `POST /agents/token/refresh`로 다시 받는다.
servers:
  - url: http://localhost:3000/api
    description: 로컬 개발 서버
components:
  securitySchemes:
    userToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 사람 토큰. `POST /auth/login`의 `data.accessToken`을 붙여넣는다.
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 에이전트 토큰. 시드 출력의 access token 또는 refresh 응답의 값을 붙여넣는다.
...
  parameters:
...
    RepoId:
      name: repoId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{REPO_API}}'
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{ORG_ID}}'
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{PROJECT_ID}}'
...
paths:
...
  /orgs:
    post:
      tags: [조직]
      summary: 조직 생성 (만든 사람이 대표)
      description: |
        **이 순간 그 사용자가 먼저 연결해 둔 에이전트의 `org_id`도 함께 채워진다.**
        한 사용자는 하나의 조직에만 속한다 — 시드 계정으로 부르면 409가 정상이다.
      security: [{ userToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 128 }
            example:
              name: Newbie Inc.
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              example:
                data: { orgId: '{{ORG_ID}}', userId: 1959e109-c60c-43f4-842f-dd18d9e7afab }
        '409':
          description: 이미 조직에 속함
          content:
            application/json:
              example:
                error: { code: ALREADY_IN_ORG, message: user already belongs to an organization }
  /orgs/{orgId}/github/repos:
    get:
      tags: [조직]
      summary: GitHub에서 접근 가능한 레포 목록 (조직 멤버)
      description: |
        레포 연결이 조직 멤버 누구나 할 수 있으므로, 그 드롭다운을 채우는 이 목록도 함께 열려 있다.

        `GITHUB_TOKEN`이 없으면 **500이 아니라 빈 배열**을 반환한다.
        GitHub 실패가 우리 기능을 멈추게 하지 않는다.
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      responses:
        '200':
          description: 레포 목록
          content:
            application/json:
              example:
                data: { repos: [] }
        '403':
          description: 대표가 아니거나 다른 조직
          content:
            application/json:
              example:
                error: { code: NOT_REPRESENTATIVE, message: this action requires the organization representative }
  /orgs/{orgId}/members:
    get:
      tags: [조직]
      summary: 조직 멤버 목록
      description: |
        GitHub collaborator 조회가 실패하면 `isCollaborator` 필드를 **생략**한다.
        `false`로 채우지 않는다 — "확인 안 됨"과 "권한 없음"은 다르다.
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      responses:
        '200':
          description: 멤버 목록
          content:
            application/json:
              example:
                data:
                  members:
                    - { userId: 1959e109-c60c-43f4-842f-dd18d9e7afab, loginId: rep, nickname: 대표, orgRole: REPRESENTATIVE }
                    - { userId: 956e2b45-f39a-4799-a256-8b03de80f4ec, loginId: be, nickname: 백엔드, orgRole: MEMBER }
        '403':
          description: 다른 조직
          content:
            application/json:
              example:
                error: { code: CROSS_ORG_ACCESS, message: cannot access another organization }
...
  /repos/{repoId}/paths:
    get:
      tags: [레포·경로]
      summary: 경로 규칙 목록 (priority 내림차순)
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/RepoId' }]
      responses:
        '200':
          description: 시드 15행 + 대표가 추가한 것
          content:
            application/json:
              example:
                data:
                  paths:
                    - { id: '{{PATH_ID}}', pathPattern: 'tests/**', ownerRole: null, access: write, actionKey: 'test:write', priority: 20, source: seed }
                    - { id: 11111111-1111-1111-1111-111111111111, pathPattern: '**', ownerRole: BACKEND, access: write, actionKey: null, priority: 10, source: seed }
        '404':
          description: 없는 레포
          content:
            application/json:
              example:
                error: { code: REPO_NOT_FOUND, message: repository not found }
    post:
      tags: [레포·경로]
      summary: 경로 규칙 추가 (대표 전용)
      description: |
        `priority`는 **manual 대역 200~299**만 받고, 안 주면 대역 최댓값 + 1이 된다(나중에 추가한 규칙이 이긴다).

        glob은 `**`, `*`, 리터럴만 허용한다. 중괄호·느낌표·물음표·대괄호는 거부된다 —
        서버와 브릿지 `.claude/settings.json`의 방언 차이로 판정이 갈리는 것을 막기 위해서다.
        **부정이 필요하면 `priority`로 푼다.**
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/RepoId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pathPattern, access]
              properties:
                pathPattern: { type: string }
                ownerRole: { type: string, enum: [FRONTEND, BACKEND], nullable: true }
                access: { type: string, enum: [write, read, denied] }
                priority: { type: integer, minimum: 200, maximum: 299 }
            example:
              pathPattern: 'api/**'
              ownerRole: BACKEND
              access: write
      responses:
        '201':
          description: 추가됨
          content:
            application/json:
              example:
                data: { path: { id: 33333333-3333-3333-3333-333333333333, pathPattern: 'api/**', ownerRole: BACKEND, access: write, priority: 200, source: manual } }
        '400':
          description: 허용하지 않는 glob 문법
          content:
            application/json:
              example:
                error: { code: INVALID_GLOB_PATTERN, message: pattern uses unsupported syntax }
        '409':
          description: 대역 안에서 priority 충돌
          content:
            application/json:
              example:
                error: { code: PATH_PRIORITY_TAKEN, message: priority 200 is already used in this repo }
  /repos/{repoId}/paths/{pathId}:
    patch:
      tags: [레포·경로]
      summary: 경로 소유권 지정 (대표 전용 — 온보딩의 핵심 산출물)
      description: |
        `actionKey`와 `priority`는 **어떤 API로도 바꿀 수 없다** — 정책 판정의 기준이라 임의로 바뀌면 재현성이 깨진다.

        priority 900 이상(조직 상한, `.env` 계열 등)은 어떤 필드도 못 바꾼다 → 409 `IMMUTABLE_ORG_CEILING`.

        ⚠️ **이 호출이 성공하면 그 레포를 쓰는 프로젝트의 `policy_hash`가 갱신된다.**
        이미 발급된 에이전트 토큰은 즉시 401 `POLICY_STALE`이 된다. 재발급하면 다시 통과한다.
      security: [{ userToken: [] }]
      parameters:
        - $ref: '#/components/parameters/RepoId'
        - name: pathId
          in: path
          required: true
          schema: { type: string, format: uuid }
          example: '{{PATH_ID}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ownerRole: { type: string, enum: [FRONTEND, BACKEND], nullable: true }
                access: { type: string, enum: [write, read, denied] }
            example:
              ownerRole: BACKEND
      responses:
        '200':
          description: 변경됨
          content:
            application/json:
              example:
                data: { path: { id: '{{PATH_ID}}', pathPattern: 'tests/**', ownerRole: BACKEND, access: write, priority: 20, source: seed } }
        '409':
          description: 조직 상한 행은 불변
          content:
            application/json:
              example:
                error: { code: IMMUTABLE_ORG_CEILING, message: organization ceiling rule cannot be changed }
...
  /orgs/{orgId}/projects:
    post:
      tags: [프로젝트]
      summary: 프로젝트 생성 (대표 전용)
      description: |
        한 트랜잭션에서 다섯 가지가 함께 만들어진다:
        프로젝트 행(`status=planning`, `started_at=NULL`) · 레포 연결 ·
        **허용 레벨의 해당 열을 복사한 `project_policies` 17행** · 조직 헌법 스냅샷 · `policy_hash`.

        판정은 이후 `action_catalog`가 아니라 **이 사본만** 본다 — 정책표가 바뀌어도 진행 중인 프로젝트는 흔들리지 않는다.
        헌법도 생성 시점 사본을 들고 간다. 조직 헌법이 나중에 바뀌어도 과거 판정의 근거가 남아야 하기 때문이다.

        **같은 레포를 두 활성 프로젝트가 쓸 수 없다**(409). 경로 소유권이 두 곳에 걸리기 때문이다.
        `status`가 `completed`·`aborted`인 프로젝트의 레포는 다시 쓸 수 있다.

        `autonomyPreset`은 zod가 아니라 서비스가 판정해서 **422**로 무엇이 허용되는지 함께 답한다.
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, autonomyPreset, pmBudgetUsd, repoIds]
              properties:
                name: { type: string, maxLength: 128 }
                autonomyPreset: { type: string, description: 'L1 | L2 | L3 | L4' }
                pmBudgetUsd:
                  type: number
                  description: PM 예산. 필수 — PM은 NOMOS 키로 돌아 비용을 우리가 부담한다.
                budgetUsd: { type: number, description: 프로젝트 전체 상한 (선택) }
                deadline: { type: string, example: '2026-12-31' }
                repoIds: { type: array, minItems: 1, items: { type: string, format: uuid } }
            example:
              name: 스터디 관리 웹앱 v2
              autonomyPreset: L2
              pmBudgetUsd: 40
              repoIds: ['{{REPO_API}}']
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              example:
                data:
                  project:
                    id: '{{PROJECT_ID}}'
                    name: 스터디 관리 웹앱 v2
                    autonomyPreset: L2
                    status: planning
                    startedAt: null
                    pmBudgetUsd: '40.00'
                    budgetUsd: null
                    # date 컬럼이지만 응답은 보낸 그대로의 YYYY-MM-DD 문자열이다
                    # (pg가 주는 Date를 toProject가 날짜 문자열로 되돌린다 — 타임존으로 밀리지 않는다).
                    deadline: '2026-12-31'
                    policyHash: bc7f3b33d92b36b25607093d7efb05cdacf73f1f0b52524046bff7b56e94442c
                    constitutionHash: 44136fa355b3678a1146ad16f7e8649e94fb4fc21fe77e8310c060f61caaff8a
                    createdBy: '{{USER_REP}}'
                  repos: [{ id: '{{REPO_API}}', orgId: '{{ORG_ID}}', fullName: acme/study-api }]
                  members: []
        '403':
          description: 대표가 아니거나 다른 조직
          content:
            application/json:
              example:
                error: { code: NOT_REPRESENTATIVE, message: this action requires the organization representative }
        '404':
          description: 없는 레포
          content:
            application/json:
              example:
                error: { code: REPO_NOT_FOUND, message: repository not found }
        '409':
          description: 활성 프로젝트가 이미 쓰는 레포
          content:
            application/json:
              example:
                error: { code: REPO_IN_ACTIVE_PROJECT, message: repository is already used by another project }
        '422':
          description: |
            허용 레벨 어휘 위반, 또는 소유 역할이 하나도 지정되지 않은 레포.

            레포 연결은 팀원도 할 수 있으므로, 소유권 지정 없이 투입되면 `owner_role`이 전부 NULL이라
            모든 역할이 레포 전체에 쓸 수 있다. 메시지에 막힌 레포 전부와 부를 API를 함께 적는다.
          content:
            application/json:
              examples:
                허용레벨:
                  value: { error: { code: INVALID_AUTONOMY_PRESET, message: 'autonomyPreset must be one of L1, L2, L3, L4' } }
                소유권미지정:
                  value: { error: { code: REPO_OWNERSHIP_NOT_SET, message: '레포의 경로 소유권을 먼저 지정하세요. PATCH /api/repos/{{REPO_API}}/paths/:pathId' } }
  /projects/{projectId}:
    get:
      tags: [프로젝트]
      summary: 프로젝트 상세 (대표 또는 배정된 에이전트의 주인)
      description: 상태·허용 레벨·`policy_hash`·레포 목록·멤버 목록을 함께 준다. DB를 직접 보지 않고 확인하기 위한 창구다.
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      responses:
        '200':
          description: 상세
          content:
            application/json:
              example:
                data:
                  project:
                    id: '{{PROJECT_ID}}'
                    status: planning
                    autonomyPreset: L2
                    policyHash: bc7f3b33d92b36b25607093d7efb05cdacf73f1f0b52524046bff7b56e94442c
                    startedAt: null
                    pmBudgetUsd: '40.00'
                  repos:
                    - { id: '{{REPO_API}}', fullName: acme/study-api }
                  members:
                    - { agentId: '{{AGENT_BE}}', agentName: be-laptop, teamRole: BACKEND }
        '403':
          description: 이 프로젝트의 멤버가 아님
          content:
            application/json:
              example:
                error: { code: NOT_PROJECT_MEMBER, message: you are not a member of this project }
        '404':
          description: 없는 프로젝트
          content:
            application/json:
              example:
                error: { code: PROJECT_NOT_FOUND, message: project not found }
  /projects/{projectId}/members:
    post:
      tags: [프로젝트]
      summary: 역할 배정 (대표 전용)
      description: |
        **배정 뒤 그 에이전트는 토큰을 재발급해야 한다.** 배정 전에 발급된 토큰에는 `project_id`가 없어
        태스크·노트 API가 403 `NOT_PROJECT_MEMBER`로 막힌다. `POST /agents/token/refresh`를 한 번 부르면 된다.
        응답의 `notice`에도 같은 안내가 들어 있다.

        프로젝트당 역할은 **하나씩만** 배정된다(`uq_project_members_role`). 실험 통제를 위해서다.
        **G1(`started_at`) 이후에는 멤버를 바꿀 수 없다** — 도중에 바뀌면 통제가 흔들린다.
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, teamRole]
              properties:
                agentId: { type: string, format: uuid }
                teamRole: { type: string, enum: [FRONTEND, BACKEND] }
            example:
              agentId: '{{AGENT_BE}}'
              teamRole: BACKEND
      responses:
        '201':
          description: 배정됨. 남은 멤버 목록과 재발급 안내를 함께 준다.
          content:
            application/json:
              example:
                data:
                  members:
                    - { agentId: '{{AGENT_BE}}', agentName: be-laptop, teamRole: BACKEND }
                  notice: 배정된 에이전트는 POST /api/agents/token/refresh로 토큰을 재발급해야 태스크·노트 API를 쓸 수 있습니다.
        '403':
          description: 대표가 아님 · 다른 조직의 에이전트 · G1 이후
          content:
            application/json:
              examples:
                남의조직에이전트:
                  value: { error: { code: AGENT_NOT_IN_ORG, message: agent does not belong to this organization } }
                시작된프로젝트:
                  value: { error: { code: PROJECT_STARTED, message: project has started; members are frozen } }
        '409':
          description: 역할이 이미 찼거나, 그 에이전트가 이미 배정됨
          content:
            application/json:
              examples:
                역할중복:
                  value: { error: { code: ROLE_ALREADY_ASSIGNED, message: BACKEND is already assigned in this project } }
                에이전트중복:
                  value: { error: { code: AGENT_ALREADY_ASSIGNED, message: this agent is already a member of the project } }
  /projects/{projectId}/members/{agentId}:
    delete:
      tags: [프로젝트]
      summary: 배정 해제 (대표 전용)
      description: |
        **역할 교체는 해제 후 재배정으로 한다.** UPDATE 경로를 두면 한 역할에 둘이 잠깐 겹칠 수 있다.
        G1(`started_at`) 이후에는 해제도 막힌다.
      security: [{ userToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: agentId
          in: path
          required: true
          schema: { type: string, format: uuid }
          example: '{{AGENT_BE}}'
      responses:
        '200':
          description: 해제됨. 남은 멤버 목록을 준다.
          content:
            application/json:
              example:
                data: { members: [] }
        '403':
          description: 대표가 아니거나 G1 이후
          content:
            application/json:
              example:
                error: { code: PROJECT_STARTED, message: project has started; members are frozen }
        '404':
          description: 배정되지 않은 에이전트
          content:
            application/json:
              example:
                error: { code: MEMBER_NOT_FOUND, message: agent is not a member of this project }
...